+++ scripts/scil_tractogram_assign_uniform_color.py
@@ -30,7 +30,7 @@ from scilpy.io.utils import (assert_inputs_exist,
                              add_overwrite_arg,
                              add_verbose_arg,
                              add_reference_arg, assert_headers_compatible)
-from scilpy.viz.utils import format_hexadecimal_color_to_rgb
+from scilpy.viz.color import format_hexadecimal_color_to_rgb
 
 
 def _build_arg_parser():
